Fix parsing of error response
Prior to this commit, only json response containing a meta attribute
were parsed properly. Instagram can also send a response body that
is the content of the meta attribute itself.

This commit adds a method that parses both cases properly and
updated the tests to take that into account

Fixes gh-58