LogHelper#prettyPrintJSONResponse lean and safer
This utility method is doing a lot of work (parsing a json, serializing it)
despite that the logger could not have debugging enable.

Also the method get call despite of the media type of the response.
There are cases where if get recent media from a user that does not exists
the 404 response is of media html. In those cases the method fails