add retry logic on response stream failure
Motivation

The instagram API, over HTTP 1.1, occasionally exhibits these
issues:

- socket reset of TCP connection
- partial JSON response stream leading to JSON parse errors

This addresses the errors by retrying the request when either a
socket exception or JSON parse error occurs.

It is worth noting that the instagram API over HTTP2 does not
seem to have these issues. It would be much more work to allow
HTTP 2, at this point, because Java does not support in their
native APIs.

Solution

Provide a configurable setting to retry the request on stream
failure.

Only one retry is attempted. If the response failed again the
errors will be propagated.

Missing

A test should be provided to check the retry logic.

retry on invalid json response

single retry; must be enabled in InstagramConfig

fix retry logic